Keep config loader in metadata module
