+++ src/commands/config/export.ts
@@ -7,8 +7,7 @@ import pc from 'picocolors';
 import { requireAuth } from '../../lib/credentials.js';
 import { handleError, getRootOpts, CLIError } from '../../lib/errors.js';
 import { stringifyConfigToml } from '../../lib/config-toml.js';
-import { configFromConfigState } from '../../lib/config-metadata.js';
-import { loadConfigState } from '../../lib/config-state.js';
+import { configFromConfigState, loadConfigState } from '../../lib/config-metadata.js';
 import { reportCliUsage } from '../../lib/skills.js';
 import { trackConfig, shutdownAnalytics } from '../../lib/analytics.js';
 import { getProjectConfig } from '../../lib/config.js';
