+++ src/commands/config/plan.ts
@@ -9,8 +9,7 @@ import { parseConfigToml } from '../../lib/config-toml.js';
 import { diffConfig } from '../../lib/config-diff.js';
 import { formatPlan } from '../../lib/config-format.js';
 import { configSupports, changePath } from '../../lib/config-capabilities.js';
-import { liveFromConfigState } from '../../lib/config-metadata.js';
-import { loadConfigState } from '../../lib/config-state.js';
+import { liveFromConfigState, loadConfigState } from '../../lib/config-metadata.js';
 import { reportCliUsage } from '../../lib/skills.js';
 import { trackConfig, shutdownAnalytics } from '../../lib/analytics.js';
 import { getProjectConfig } from '../../lib/config.js';
